+++ lib/tiktok.rb
@@ -1,12 +1,14 @@
 # frozen_string_literal: true
 
 require "net/http"
+require "uri"
 require "json"
 
 require_relative "tiktok/version"
 require_relative "tiktok/errors"
 require_relative "tiktok/handle"
 require_relative "tiktok/url"
+require_relative "tiktok/short_url"
 require_relative "tiktok/sound"
 require_relative "tiktok/account"
 require_relative "tiktok/video"